Rename updater
